refactor: move function in new file
